Refactor & reorganize
