+++ MVC.Budget.Ananttuli/Budget/Views/_ViewImports.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Budget
-@using Budget.Models
+@using Budget.ViewModels
 @using Budget.CategoriesModule.Models
 @using Budget.TransactionsModule.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
